+++ package.js
@@ -2,14 +2,14 @@
 Package.describe({
   name: 'space:accounts',
   summary: 'Accounts module for Space applications',
-  version: '0.1.3',
+  version: '0.2.0',
   git: 'https://github.com/meteor-space/accounts.git',
   documentation: 'README.md'
 });
 
 Package.onUse(function(api) {
 
-  api.versionsFrom('METEOR@1.0');
+  api.versionsFrom('1.2.0.1');
 
   api.use([
     'coffeescript',
@@ -19,40 +19,17 @@ Package.onUse(function(api) {
     'check',
     'ecmascript',
     'underscore',
-    'space:vo-user@0.2.1',
-    'space:event-sourcing@2.1.0'
+    'space:base@4.1.1',
+    'space:vo-user@0.3.0',
+    'space:messaging@3.1.1'
   ]);
 
   api.addFiles([
-    'source/server/module.coffee',
-    'source/server/commands.coffee',
-    'source/server/events.coffee',
-    // SERVICES
-    'source/server/services/meteor-users-service.js',
-    // REGISTRATION
-    'source/server/registration/registration.js',
-    'source/server/registration/registration-router.js',
-    // USERS
-    'source/server/accounts/account.coffee',
-    'source/server/accounts/accounts-router.coffee'
-  ], 'server');
-
-});
-
-Package.onTest(function(api) {
-
-  api.use([
-    'coffeescript',
-    'mongo',
-    'underscore',
-    'space:testing@2.0.1',
-    'space:accounts',
-    'space:vo-user@0.2.0',
-    'practicalmeteor:munit@2.1.5'
-  ]);
-
-  api.addFiles([
-    'tests/accounts/accounts.test.coffee'
+    'source/server/module.js',
+    'source/server/events.js',
+    'source/server/commands.js',
+    'source/server/services/user-creation-service.js',
+    'source/server/publishers/login-publisher.js'
   ], 'server');
 
 });