+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 Changelog
 =========
 
+## 0.2.0
+Version bump space:messaging and space:user dependencies
+
 ## 0.1.0
+Initial release of Accounts module for Space applications